rename `encryption.py` to `encryption_info.py`
